package/zstd: rework build and install
Create libzstd.pc in the build step instead of triggering a
lazy build during installation when it is missing.

Move PREFIX=/usr to ZSTD_OPTS. It is needed for building libzstd.pc
during the build step; for consistency, make sure it's available in all
steps.

Attach '-release' to the targets, since the default is to
build lib-release and zstd-release. Its only effect is to clear
DEBUGFLAGS (which are just warning flags).

Signed-off-by: Norbert Lange <nolange79@gmail.com>
[Arnout: add check-package disable]
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>