package/mtd: ubihealthd needs kernel 3.17+
ubihealthd requires getrandom(2) that was introduced in kernel version
3.17. ubihealthd does not build when getrandom(2) is not detected, so
the following installation step fails.

Technically the dependency should also be on glibc version 2.25+. But we
have no way to depend on glibc versions of external toolchains.
Toolchain built with kernel headers older than 3.17 can build
ubihealthd, but it will fail at run-time. So this is a pretty close
approximation of the actual dependency.

Fixes:
http://autobuild.buildroot.net/results/2d42b0a626367e4051d0e2aadcce39e974fe09d4/
http://autobuild.buildroot.net/results/a2b6dbf707275e3f8262479c0650cfc7cb9abc8d/

Cc: Matt Weber <matthew.weber@collins.com>
Signed-off-by: Baruch Siach <baruch@tkos.co.il>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>