board/acmesystems/{aria, arietta}-g25: fix genimage.cfg
On the Aria and the Arietta AT91Bootstrap builds, the file name of the
bootloader embeds its version number, and the genimage configuration
needs this filename in order to build the boot filesystem image. Commit
b31d6a724aa780d2d76c64c184f66aa8a9884f75 bumped the AT91Bootstrap
version of all acmesystems' boards but failed to update genimage.cfg
accordingly, which broke the builds. The Acqua board is not affected
by this issue.

Update the affected genimage.cfg with the correct filenames.

Fixes:
https://gitlab.com/buildroot.org/buildroot/-/jobs/1515521690
https://gitlab.com/buildroot.org/buildroot/-/jobs/1515521691
https://gitlab.com/buildroot.org/buildroot/-/jobs/1515521692
https://gitlab.com/buildroot.org/buildroot/-/jobs/1515521694

Signed-off-by: Edgar Bonet <bonet@grenoble.cnrs.fr>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>