package/python-augeas: switch from "depends on" to "select" for augeas dep
During the review of the lua-augeas package, it was pointed out that a
"select BR2_PACKAGE_AUGEAS" should be used instead of a "depends on
BR2_PACKAGE_AUGEAS". But, the same review pointed that python-augeas
used a "depends on BR2_PACKAGE_AUGEAS".

In order to make things consistent, let's switch to "select
BR2_PACKAGE_AUGEAS" as well for python-augeas.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>